+++ include/matrix.h
@@ -6,15 +6,15 @@
 /*   By: roberto <roberto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/22 18:24:15 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/08 18:11:50 by roberto          ###   ########.fr       */
+/*   Updated: 2023/03/08 21:28:19 by roberto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MATRIX_H
 # define MATRIX_H
 
-typedef float	t_matrix4[4][4];
-typedef float	t_matrix3[3][3];
-typedef float	t_matrix2[2][2];
+typedef double	t_matrix4[4][4];
+typedef double	t_matrix3[3][3];
+typedef double	t_matrix2[2][2];
 
 #endif
