+++ src/intersection/intersect_caps.c
@@ -6,13 +6,13 @@
 /*   By: roberto <roberto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/28 12:24:28 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/08 21:10:41 by roberto          ###   ########.fr       */
+/*   Updated: 2023/03/08 21:28:22 by roberto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <header.h>
 
-static int	check_cap(t_ray ray, float t);
+static int	check_cap(t_ray ray, double t);
 
 /* ************************************************************************** */
 /*                                                                            */
@@ -22,7 +22,7 @@ static int	check_cap(t_ray ray, float t);
 /* ************************************************************************** */
 void	intersect_caps(t_ray ray, t_object *object, t_intersection *ret)
 {
-	float	t;
+	double	t;
 
 	if (!object->cyl_closed || (ray.vector.y >= -0.1 && ray.vector.y <= 0.1))
 		return ;
@@ -44,10 +44,10 @@ void	intersect_caps(t_ray ray, t_object *object, t_intersection *ret)
 	}
 }
 
-static int	check_cap(t_ray ray, float t)
+static int	check_cap(t_ray ray, double t)
 {
-	float	x;
-	float	z;
+	double	x;
+	double	z;
 
 	x = ray.point.x + t * ray.vector.x;
 	z = ray.point.z + t * ray.vector.z;
