+++ src/tuple/init_vector.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   init_vector.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: roberto <roberto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/18 17:05:37 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/02/18 17:09:21 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/03/08 21:28:32 by roberto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <header.h>
 
-void	init_vector(t_tuple *tuple, float x, float y, float z)
+void	init_vector(t_tuple *tuple, double x, double y, double z)
 {
 	tuple->x = x;
 	tuple->y = y;
