+++ src/world/give_camera.c
@@ -6,7 +6,7 @@
 /*   By: roberto <roberto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/24 14:42:14 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/08 06:40:59 by roberto          ###   ########.fr       */
+/*   Updated: 2023/03/08 21:28:37 by roberto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,11 @@
 /*   agrandissant une zone plus petite de la scène.							  */
 /*                                                                            */
 /* ************************************************************************** */
-t_camera	give_camera(float hsize, float vsize, float field_of_view)
+t_camera	give_camera(double hsize, double vsize, double field_of_view)
 {
 	t_camera	ret;
-	float		half_view;
-	float		aspect;
+	double		half_view;
+	double		aspect;
 
 	ret.hsize = hsize;
 	ret.vsize = vsize;
